+++ AWSIoTPythonSDK/__init__.py
@@ -1,6 +1,6 @@
 import os
 import sys
 
-__version__ = "1.1.1"
+__version__ = "1.1.2"