+++ samples/basicPubSub/basicPubSub.py
@@ -19,7 +19,7 @@ from AWSIoTPythonSDK.MQTTLib import AWSIoTMQTTClient
 import sys
 import logging
 import time
-import getopt
+import argparse
 
 # Custom MQTT message callback
 def customCallback(client, userdata, message):
@@ -29,78 +29,32 @@ def customCallback(client, userdata, message):
 	print(message.topic)
 	print("--------------\n\n")
 
-# Usage
-usageInfo = """Usage:
-
-Use certificate based mutual authentication:
-python basicPubSub.py -e <endpoint> -r <rootCAFilePath> -c <certFilePath> -k <privateKeyFilePath>
-
-Use MQTT over WebSocket:
-python basicPubSub.py -e <endpoint> -r <rootCAFilePath> -w
-
-Type "python basicPubSub.py -h" for available options.
-"""
-# Help info
-helpInfo = """-e, --endpoint
-	Your AWS IoT custom endpoint
--r, --rootCA
-	Root CA file path
--c, --cert
-	Certificate file path
--k, --key
-	Private key file path
--w, --websocket
-	Use MQTT over WebSocket
--h, --help
-	Help information
-
-
-"""
-
 # Read in command-line parameters
-useWebsocket = False
-host = ""
-rootCAPath = ""
-certificatePath = ""
-privateKeyPath = ""
-try:
-	opts, args = getopt.getopt(sys.argv[1:], "hwe:k:c:r:", ["help", "endpoint=", "key=","cert=","rootCA=", "websocket"])
-	if len(opts) == 0:
-		raise getopt.GetoptError("No input parameters!")
-	for opt, arg in opts:
-		if opt in ("-h", "--help"):
-			print(helpInfo)
-			exit(0)
-		if opt in ("-e", "--endpoint"):
-			host = arg
-		if opt in ("-r", "--rootCA"):
-			rootCAPath = arg
-		if opt in ("-c", "--cert"):
-			certificatePath = arg
-		if opt in ("-k", "--key"):
-			privateKeyPath = arg
-		if opt in ("-w", "--websocket"):
-			useWebsocket = True
-except getopt.GetoptError:
-	print(usageInfo)
-	exit(1)
+parser = argparse.ArgumentParser()
+parser.add_argument("-e", "--endpoint", action="store", required=True, dest="host", help="Your AWS IoT custom endpoint")
+parser.add_argument("-r", "--rootCA", action="store", required=True, dest="rootCAPath", help="Root CA file path")
+parser.add_argument("-c", "--cert", action="store", dest="certificatePath", help="Certificate file path")
+parser.add_argument("-k", "--key", action="store", dest="privateKeyPath", help="Private key file path")
+parser.add_argument("-w", "--websocket", action="store_true", dest="useWebsocket", default=False,
+                    help="Use MQTT over WebSocket")
+parser.add_argument("-id", "--clientId", action="store", dest="clientId", default="basicPubSub", help="Targeted client id")
+parser.add_argument("-t", "--topic", action="store", dest="topic", default="sdk/test/Python", help="Targeted topic")
 
-# Missing configuration notification
-missingConfiguration = False
-if not host:
-	print("Missing '-e' or '--endpoint'")
-	missingConfiguration = True
-if not rootCAPath:
-	print("Missing '-r' or '--rootCA'")
-	missingConfiguration = True
-if not useWebsocket:
-	if not certificatePath:
-		print("Missing '-c' or '--cert'")
-		missingConfiguration = True
-	if not privateKeyPath:
-		print("Missing '-k' or '--key'")
-		missingConfiguration = True
-if missingConfiguration:
+args = parser.parse_args()
+host = args.host
+rootCAPath = args.rootCAPath
+certificatePath = args.certificatePath
+privateKeyPath = args.privateKeyPath
+useWebsocket = args.useWebsocket
+clientId = args.clientId
+topic = args.topic
+
+if args.useWebsocket and args.certificatePath and args.privateKeyPath:
+	parser.error("X.509 cert authentication and WebSocket are mutual exclusive. Please pick one.")
+	exit(2)
+
+if not args.useWebsocket and (not args.certificatePath or not args.privateKeyPath):
+	parser.error("Missing credentials for authentication.")
 	exit(2)
 
 # Configure logging
@@ -114,11 +68,11 @@ logger.addHandler(streamHandler)
 # Init AWSIoTMQTTClient
 myAWSIoTMQTTClient = None
 if useWebsocket:
-	myAWSIoTMQTTClient = AWSIoTMQTTClient("basicPubSub", useWebsocket=True)
+	myAWSIoTMQTTClient = AWSIoTMQTTClient(clientId, useWebsocket=True)
 	myAWSIoTMQTTClient.configureEndpoint(host, 443)
 	myAWSIoTMQTTClient.configureCredentials(rootCAPath)
 else:
-	myAWSIoTMQTTClient = AWSIoTMQTTClient("basicPubSub")
+	myAWSIoTMQTTClient = AWSIoTMQTTClient(clientId)
 	myAWSIoTMQTTClient.configureEndpoint(host, 8883)
 	myAWSIoTMQTTClient.configureCredentials(rootCAPath, privateKeyPath, certificatePath)
 
@@ -131,12 +85,12 @@ myAWSIoTMQTTClient.configureMQTTOperationTimeout(5)  # 5 sec
 
 # Connect and subscribe to AWS IoT
 myAWSIoTMQTTClient.connect()
-myAWSIoTMQTTClient.subscribe("sdk/test/Python", 1, customCallback)
+myAWSIoTMQTTClient.subscribe(topic, 1, customCallback)
 time.sleep(2)
 
 # Publish to the same topic in a loop forever
 loopCount = 0
 while True:
-	myAWSIoTMQTTClient.publish("sdk/test/Python", "New Message " + str(loopCount), 1)
+	myAWSIoTMQTTClient.publish(topic, "New Message " + str(loopCount), 1)
 	loopCount += 1
 	time.sleep(1)