+++ samples/basicShadow/basicShadowUpdater.py
@@ -20,7 +20,8 @@ import sys
 import logging
 import time
 import json
-import getopt
+import argparse
+
 
 # Shadow JSON schema:
 #
@@ -31,108 +32,60 @@ import getopt
 #			"property":<INT VALUE>
 #		}
 #	}
-#}
+# }
 
 # Custom Shadow callback
 def customShadowCallback_Update(payload, responseStatus, token):
-	# payload is a JSON string ready to be parsed using json.loads(...)
-	# in both Py2.x and Py3.x
-	if responseStatus == "timeout":
-		print("Update request " + token + " time out!")
-	if responseStatus == "accepted":
-		payloadDict = json.loads(payload)
-		print("~~~~~~~~~~~~~~~~~~~~~~~")
-		print("Update request with token: " + token + " accepted!")
-		print("property: " + str(payloadDict["state"]["desired"]["property"]))
-		print("~~~~~~~~~~~~~~~~~~~~~~~\n\n")
-	if responseStatus == "rejected":
-		print("Update request " + token + " rejected!")
+    # payload is a JSON string ready to be parsed using json.loads(...)
+    # in both Py2.x and Py3.x
+    if responseStatus == "timeout":
+        print("Update request " + token + " time out!")
+    if responseStatus == "accepted":
+        payloadDict = json.loads(payload)
+        print("~~~~~~~~~~~~~~~~~~~~~~~")
+        print("Update request with token: " + token + " accepted!")
+        print("property: " + str(payloadDict["state"]["desired"]["property"]))
+        print("~~~~~~~~~~~~~~~~~~~~~~~\n\n")
+    if responseStatus == "rejected":
+        print("Update request " + token + " rejected!")
 
 def customShadowCallback_Delete(payload, responseStatus, token):
-	if responseStatus == "timeout":
-		print("Delete request " + token + " time out!")
-	if responseStatus == "accepted":
-		print("~~~~~~~~~~~~~~~~~~~~~~~")
-		print("Delete request with token: " + token + " accepted!")
-		print("~~~~~~~~~~~~~~~~~~~~~~~\n\n")
-	if responseStatus == "rejected":
-		print("Delete request " + token + " rejected!")
-
-# Usage
-usageInfo = """Usage:
-
-Use certificate based mutual authentication:
-python basicShadowUpdater.py -e <endpoint> -r <rootCAFilePath> -c <certFilePath> -k <privateKeyFilePath>
-
-Use MQTT over WebSocket:
-python basicShadowUpdater.py -e <endpoint> -r <rootCAFilePath> -w
-
-Type "python basicShadowUpdater.py -h" for available options.
-
-
-"""
-# Help info
-helpInfo = """-e, --endpoint
-	Your AWS IoT custom endpoint
--r, --rootCA
-	Root CA file path
--c, --cert
-	Certificate file path
--k, --key
-	Private key file path
--w, --websocket
-	Use MQTT over WebSocket
--h, --help
-	Help information
-
-
-"""
+    if responseStatus == "timeout":
+        print("Delete request " + token + " time out!")
+    if responseStatus == "accepted":
+        print("~~~~~~~~~~~~~~~~~~~~~~~")
+        print("Delete request with token: " + token + " accepted!")
+        print("~~~~~~~~~~~~~~~~~~~~~~~\n\n")
+    if responseStatus == "rejected":
+        print("Delete request " + token + " rejected!")
 
 # Read in command-line parameters
-useWebsocket = False
-host = ""
-rootCAPath = ""
-certificatePath = ""
-privateKeyPath = ""
-try:
-	opts, args = getopt.getopt(sys.argv[1:], "hwe:k:c:r:", ["help", "endpoint=", "key=","cert=","rootCA=", "websocket"])
-	if len(opts) == 0:
-		raise getopt.GetoptError("No input parameters!")
-	for opt, arg in opts:
-		if opt in ("-h", "--help"):
-			print(helpInfo)
-			exit(0)
-		if opt in ("-e", "--endpoint"):
-			host = arg
-		if opt in ("-r", "--rootCA"):
-			rootCAPath = arg
-		if opt in ("-c", "--cert"):
-			certificatePath = arg
-		if opt in ("-k", "--key"):
-			privateKeyPath = arg
-		if opt in ("-w", "--websocket"):
-			useWebsocket = True
-except getopt.GetoptError:
-	print(usageInfo)
-	exit(1)
+parser = argparse.ArgumentParser()
+parser.add_argument("-e", "--endpoint", action="store", required=True, dest="host", help="Your AWS IoT custom endpoint")
+parser.add_argument("-r", "--rootCA", action="store", required=True, dest="rootCAPath", help="Root CA file path")
+parser.add_argument("-c", "--cert", action="store", dest="certificatePath", help="Certificate file path")
+parser.add_argument("-k", "--key", action="store", dest="privateKeyPath", help="Private key file path")
+parser.add_argument("-w", "--websocket", action="store_true", dest="useWebsocket", default=False,
+                    help="Use MQTT over WebSocket")
+parser.add_argument("-n", "--thingName", action="store", dest="thingName", default="Bot", help="Targeted thing name")
+parser.add_argument("-id", "--clientId", action="store", dest="clientId", default="basicShadowUpdater", help="Targeted client id")
 
-# Missing configuration notification
-missingConfiguration = False
-if not host:
-	print("Missing '-e' or '--endpoint'")
-	missingConfiguration = True
-if not rootCAPath:
-	print("Missing '-r' or '--rootCA'")
-	missingConfiguration = True
-if not useWebsocket:
-	if not certificatePath:
-		print("Missing '-c' or '--cert'")
-		missingConfiguration = True
-	if not privateKeyPath:
-		print("Missing '-k' or '--key'")
-		missingConfiguration = True
-if missingConfiguration:
-	exit(2)
+args = parser.parse_args()
+host = args.host
+rootCAPath = args.rootCAPath
+certificatePath = args.certificatePath
+privateKeyPath = args.privateKeyPath
+useWebsocket = args.useWebsocket
+thingName = args.thingName
+clientId = args.clientId
+
+if args.useWebsocket and args.certificatePath and args.privateKeyPath:
+    parser.error("X.509 cert authentication and WebSocket are mutual exclusive. Please pick one.")
+    exit(2)
+
+if not args.useWebsocket and (not args.certificatePath or not args.privateKeyPath):
+    parser.error("Missing credentials for authentication.")
+    exit(2)
 
 # Configure logging
 logger = logging.getLogger("AWSIoTPythonSDK.core")
@@ -145,13 +98,13 @@ logger.addHandler(streamHandler)
 # Init AWSIoTMQTTShadowClient
 myAWSIoTMQTTShadowClient = None
 if useWebsocket:
-	myAWSIoTMQTTShadowClient = AWSIoTMQTTShadowClient("basicShadowUpdater", useWebsocket=True)
-	myAWSIoTMQTTShadowClient.configureEndpoint(host, 443)
-	myAWSIoTMQTTShadowClient.configureCredentials(rootCAPath)
+    myAWSIoTMQTTShadowClient = AWSIoTMQTTShadowClient(clientId, useWebsocket=True)
+    myAWSIoTMQTTShadowClient.configureEndpoint(host, 443)
+    myAWSIoTMQTTShadowClient.configureCredentials(rootCAPath)
 else:
-	myAWSIoTMQTTShadowClient = AWSIoTMQTTShadowClient("basicShadowUpdater")
-	myAWSIoTMQTTShadowClient.configureEndpoint(host, 8883)
-	myAWSIoTMQTTShadowClient.configureCredentials(rootCAPath, privateKeyPath, certificatePath)
+    myAWSIoTMQTTShadowClient = AWSIoTMQTTShadowClient(clientId)
+    myAWSIoTMQTTShadowClient.configureEndpoint(host, 8883)
+    myAWSIoTMQTTShadowClient.configureCredentials(rootCAPath, privateKeyPath, certificatePath)
 
 # AWSIoTMQTTShadowClient configuration
 myAWSIoTMQTTShadowClient.configureAutoReconnectBackoffTime(1, 32, 20)
@@ -162,15 +115,15 @@ myAWSIoTMQTTShadowClient.configureMQTTOperationTimeout(5)  # 5 sec
 myAWSIoTMQTTShadowClient.connect()
 
 # Create a deviceShadow with persistent subscription
-Bot = myAWSIoTMQTTShadowClient.createShadowHandlerWithName("Bot", True)
+deviceShadowHandler = myAWSIoTMQTTShadowClient.createShadowHandlerWithName(thingName, True)
 
 # Delete shadow JSON doc
-Bot.shadowDelete(customShadowCallback_Delete, 5)
+deviceShadowHandler.shadowDelete(customShadowCallback_Delete, 5)
 
 # Update shadow in a loop
 loopCount = 0
 while True:
-	JSONPayload = '{"state":{"desired":{"property":' + str(loopCount) + '}}}'
-	Bot.shadowUpdate(JSONPayload, customShadowCallback_Update, 5)
-	loopCount += 1
-	time.sleep(1)
+    JSONPayload = '{"state":{"desired":{"property":' + str(loopCount) + '}}}'
+    deviceShadowHandler.shadowUpdate(JSONPayload, customShadowCallback_Update, 5)
+    loopCount += 1
+    time.sleep(1)